+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/processor_strategies/base.py
@@ -46,7 +46,6 @@ class ProcessingExecutionStrategy(WithConstructionConfig, ABC):
     async def process(
         self,
         documents: Iterable[DocumentMeta | Document | Source],
-        embedder: Embedder,
         vector_store: VectorStore,
         processor_router: DocumentProcessorRouter,
         processor_overwrite: BaseProvider | None = None,
@@ -56,7 +55,6 @@ class ProcessingExecutionStrategy(WithConstructionConfig, ABC):
 
         Args:
             documents: The documents to process.
-            embedder: The embedder to produce chunk embeddings.
             vector_store: The vector store to store document chunks.
             processor_router: The document processor router to use.
             processor_overwrite: Forces the use of a specific processor, instead of the one provided by the router.
@@ -95,10 +93,10 @@ class ProcessingExecutionStrategy(WithConstructionConfig, ABC):
     @staticmethod
     async def _remove_entries_with_same_sources(elements: list[Element], vector_store: VectorStore) -> None:
         """
-        Remove entries from the vector store whose source id is present in the elements' metadata.
+        Remove entries from the vector store whose source id is present in the elements metadata.
 
         Args:
-            elements: List of elements whose source ids will be checked and removed from the vector store if present.
+            elements: The list of elements whose source ids will be removed from the vector store.
             vector_store: The vector store to store document chunks.
         """
         unique_source_ids = {element.document_meta.source.id for element in elements}
@@ -113,40 +111,12 @@ class ProcessingExecutionStrategy(WithConstructionConfig, ABC):
             await vector_store.remove(ids_to_delete)
 
     @staticmethod
-    async def _insert_elements(elements: list[Element], embedder: Embedder, vector_store: VectorStore) -> None:
+    async def _insert_elements(elements: list[Element], vector_store: VectorStore) -> None:
         """
         Insert Elements into the vector store.
 
         Args:
-            elements: The list of Elements to insert.
-            embedder: The embedder to produce chunk embeddings.
+            elements: The list of elements to insert.
             vector_store: The vector store to store document chunks.
         """
-        elements_with_text = [element for element in elements if element.key]
-        images_with_text = [element for element in elements_with_text if isinstance(element, ImageElement)]
-        vectors = await embedder.embed_text([element.key for element in elements_with_text])  # type: ignore
-
-        image_elements = [element for element in elements if isinstance(element, ImageElement)]
-
-        entries = [
-            element.to_vector_db_entry(vector, EmbeddingType.TEXT)
-            for element, vector in zip(elements_with_text, vectors, strict=False)
-        ]
-        not_embedded_image_elements = [
-            image_element for image_element in image_elements if image_element not in images_with_text
-        ]
-
-        if image_elements and embedder.image_support():
-            image_vectors = await embedder.embed_image([element.image_bytes for element in image_elements])
-            entries.extend(
-                [
-                    element.to_vector_db_entry(vector, EmbeddingType.IMAGE)
-                    for element, vector in zip(image_elements, image_vectors, strict=False)
-                ]
-            )
-            not_embedded_image_elements = []
-
-        for image_element in not_embedded_image_elements:
-            warnings.warn(f"Image: {image_element.id} could not be embedded")
-
-        await vector_store.store(entries)
+        await vector_store.store([element.to_vector_db_entry() for element in elements])