+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/processor_strategies/distributed.py
@@ -36,7 +36,6 @@ class DistributedProcessing(ProcessingExecutionStrategy):
     async def process(  # noqa: PLR6301
         self,
         documents: Iterable[DocumentMeta | Document | Source],
-        embedder: Embedder,
         vector_store: VectorStore,
         processor_router: DocumentProcessorRouter,
         processor_overwrite: BaseProvider | None = None,
@@ -46,7 +45,6 @@ class DistributedProcessing(ProcessingExecutionStrategy):
 
         Args:
             documents: The documents to process.
-            embedder: The embedder to produce chunk embeddings.
             vector_store: The vector store to store document chunks.
             processor_router: The document processor router to use.
             processor_overwrite: Forces the use of a specific processor, instead of the one provided by the router.