+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/processor_strategies/sequential.py
@@ -18,7 +18,6 @@ class SequentialProcessing(ProcessingExecutionStrategy):
     async def process(
         self,
         documents: Iterable[DocumentMeta | Document | Source],
-        embedder: Embedder,
         vector_store: VectorStore,
         processor_router: DocumentProcessorRouter,
         processor_overwrite: BaseProvider | None = None,
@@ -28,7 +27,6 @@ class SequentialProcessing(ProcessingExecutionStrategy):
 
         Args:
             documents: The documents to process.
-            embedder: The embedder to produce chunk embeddings.
             vector_store: The vector store to store document chunks.
             processor_router: The document processor router to use.
             processor_overwrite: Forces the use of a specific processor, instead of the one provided by the router.
@@ -50,7 +48,6 @@ class SequentialProcessing(ProcessingExecutionStrategy):
                 )
                 await self._insert_elements(
                     elements=elements,
-                    embedder=embedder,
                     vector_store=vector_store,
                 )
             except Exception: